Rename generator back to document and document to manifest
