+++ lib/resume_generator/cli/application.rb
@@ -50,7 +50,7 @@ module ResumeGenerator
         require 'prawn'
         require 'prawn/table'
         inform_start_of_resume_generation
-        Resume::Generator.start(self)
+        Resume::Document.generate(self)
         inform_of_successful_resume_generation
       end
